requirements.txt
fastapi==0.116.1
maxminddb==2.6.2
passlib==1.7.4
pydantic==2.11.7
python_jose==3.5.0
PyYAML==6.0.2
PyYAML==6.0.2
redis==6.4.0
Requests==2.32.5
SQLAlchemy==2.0.43
uvicorn==0.35.0
